+++ changelog.js
@@ -4,6 +4,6 @@ var Fs = require('fs');
 return Changelog.generate({
   patch: true,
   repoUrl: 'https://github.com/soulsam480/pdf-donkey',
-}).then(function (changelog) {
+}).then((changelog) => {
   Fs.writeFileSync('./CHANGELOG.md', changelog);
 });