+++ packages/api/src/controllers/templateController.ts
@@ -1,18 +1,17 @@
 import {
   Body,
   Controller,
-  CurrentUser,
   Delete,
   Get,
   HttpError,
   Param,
   Post,
   Put,
+  Req,
   UseBefore,
 } from 'routing-controllers';
 import { TemplateEntity } from 'src/entities/template';
-import { User } from 'src/entities/user';
-import { authMiddleware } from 'src/middlewares/auth.middleware';
+import { authMiddleware, RequestWithUser } from 'src/middlewares/auth.middleware';
 import { getRepository } from 'typeorm';
 import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';
 
@@ -21,12 +20,12 @@ import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity
 export class userController {
   private readonly templateRepo = getRepository(TemplateEntity);
   @Get('/')
-  async getAll(@CurrentUser() user: User) {
+  async getAll(@Req() { userId }: RequestWithUser) {
     try {
       return await this.templateRepo
         .find({
           where: {
-            user: user.id,
+            user: userId,
           },
         })
         .then((res) =>
@@ -38,10 +37,10 @@ export class userController {
   }
 
   @Get('/:id')
-  async getById(@Param('id') id: string, @CurrentUser() user: User) {
+  async getById(@Param('id') id: string, @Req() { userId }: RequestWithUser) {
     try {
       const template = await this.templateRepo.findOne({
-        where: { id: id, user: user.id },
+        where: { id: id, user: userId },
       });
       return {
         ...template,
@@ -54,14 +53,14 @@ export class userController {
   }
 
   @Post('/')
-  async createTemplate(@Body() template: TemplateEntity, @CurrentUser() user: User) {
+  async createTemplate(@Body() template: TemplateEntity, @Req() { userId }: RequestWithUser) {
     try {
       return this.templateRepo
         .create({
           ...template,
           data: template.data ? JSON.stringify(template.data) : '{}',
           meta: template.data ? JSON.stringify(template.meta) : '{}',
-          user: { id: user.id },
+          user: { id: userId },
         })
         .save();
     } catch (error) {
@@ -73,11 +72,11 @@ export class userController {
   async updateById(
     @Param('id') id: string,
     @Body() template: QueryDeepPartialEntity<TemplateEntity>,
-    @CurrentUser() user: User,
+    @Req() { userId }: RequestWithUser,
   ) {
     try {
       return await this.templateRepo.update(
-        { id: id, user: user },
+        { id: id, user: { id: userId } },
         { ...template, data: JSON.stringify(template.data), meta: JSON.stringify(template.meta) },
       );
     } catch (error) {
@@ -86,9 +85,9 @@ export class userController {
   }
 
   @Delete('/:id')
-  async deleteById(@Param('id') id: string, @CurrentUser() user: User) {
+  async deleteById(@Param('id') id: string, @Req() { userId }: RequestWithUser) {
     try {
-      return await this.templateRepo.delete({ id: id, user: user });
+      return await this.templateRepo.delete({ id: id, user: { id: userId } });
     } catch (error) {
       return new HttpError(500, error);
     }